requirements.txt
flask-sqlalchemy
flask-bootstrap
flask-migrate
flask-script